New Simutrans-Experimental specific save game version. Now saves data about (1) reversed state; (2) vehicle replacing; (3) origins and timings; (4) weight limits; and (5) settings.

+++ simcity.h
@@ -169,7 +169,7 @@ private:
 
 	slist_tpl<stadtauto_t *> current_cars;
 
-	int route_result;
+	uint8 route_result;
 
 public:
 	/**
